requirements.txt
Flask
Flask-MySQLdb
python-dotenv
Pillow